Implement some first Cpp tests after fixing compilation

+++ src/_pylimer_tools/entities/Universe.h
@@ -17,8 +17,8 @@ namespace pylimer_tools
     public:
       Universe(const double Lx, const double Ly, const double Lz);
       void setBoxLengths(const double Lx, const double Ly, const double Lz);
-      void addAtoms(const int NNewAtoms, std::vector<long int> ids, std::vector<int> types, std::vector<double> x, std::vector<double> y, std::vector<double> z, std::vector<int> nx, std::vector<int> ny, std::vector<int> nz);
-      void addBonds(const int NNewBonds, std::vector<long int> from, std::vector<long int> to);
+      void addAtoms(const size_t NNewAtoms, std::vector<long int> ids, std::vector<int> types, std::vector<double> x, std::vector<double> y, std::vector<double> z, std::vector<int> nx, std::vector<int> ny, std::vector<int> nz);
+      void addBonds(const size_t NNewBonds, std::vector<long int> from, std::vector<long int> to);
       std::vector<Molecule> getMolecules(const int atomTypeToOmit);
       std::vector<Molecule> getChainsWithCrosslinker(const int crosslinkerType);
       std::map<int, int> determineFunctionalityPerType();
